fix: refactor for future reuse
