+++ packages/xml-views-definition/src/utils/file.ts
@@ -21,19 +21,18 @@ export async function pathExists(filePath: string): Promise<boolean> {
 
 /**
  * Build absolute file uri based on manifest path and value without namespace.
- * Following extension are checked:
- * - `.controller.js`
- * - `.controller.ts`
  *
  * @param namespace app id or app namespace
  * @param value value with namespace
  * @param manifestPath path to manifest
+ * @param exts file extension
  * @returns file uri or undefined
  */
 export async function buildFileUri(
   namespace: string,
   value: string,
-  manifestPath: string
+  manifestPath: string,
+  exts: string[]
 ): Promise<string | undefined> {
   /* istanbul ignore next */
   const parts = value.split(namespace) ?? [];
@@ -42,14 +41,10 @@ export async function buildFileUri(
   const withoutNSParts = valueWithoutNS.split(".") ?? [];
   const absolutePath = join(dirname(manifestPath), ...withoutNSParts);
 
-  for (const ext of ["js", "ts"]) {
-    const filePath = `${absolutePath}.controller.${ext}`;
+  for (const ext of exts) {
+    const filePath = `${absolutePath}${ext}`;
     if (await pathExists(filePath)) {
       return filePath;
-    }
-    const filePathWithoutCtr = `${absolutePath}.${ext}`;
-    if (await pathExists(filePathWithoutCtr)) {
-      return filePathWithoutCtr;
     }
   }
   return;
